+++ CHANGES.rst
@@ -11,6 +11,7 @@ Added:
 - ``pre_flush`` and ``post_flush`` signals
 - Support for Django 1.11 and Python 3.6
 - Support of proxy models using proxy groups
+- Support of options in django's Meta
 
 
 v0.4 (09-12-2015)